Modify UsersImport to handle password field
Moved the handling of the password property in the UsersImport.php file to ensure the user creation is successful. Now, the password property is set to null as part of other collected data, right before the user creation, instead of beforehand. It could prevent potential issues in the user import process.