Implement user spreadsheet download functionality
Added the capability for the admin to download all users as a spreadsheet. This is achieved by creating new controller files and methods for exporting users. Modifications were also made in the relevant Vue file to include a button for this feature. On the other hand, the import user function was streamlined by removing the previous downloadable 'template' Excel file from the codebase, reducing potential issues and making the code cleaner.