Add shift capacity validation during shift reservation
Extracted logic of shift capacity validation to separate action. Now, the capacity of a shift needs to be validated not only during shift assignment but also when moving users to new shifts or changing shift reservations. This enhances the robustness of the shift management system against shift overflows. In case the shift is already at its maximum capacity, a specific ShiftAvailabilityException will be thrown.