Refactor assertions in ShiftsTest for fluent syntax
The code has been updated to use a fluent syntax for HTTP test assertions in the ShiftsTest class. This syntax not only makes the code cleaner and easier to read but also simplifies the debugging process as it reduces the quantity of local variables. Also added in test_not_enabled_user_cannot_see_shifts