Update shift reservation validation rule
The validation rule for shift reservations in ToggleShiftReservationControllerRules has been updated. A conditional check has been added to only enforce uniqueness of 'shift_user', 'shift_id' when a reservation is in progress. This is to ensure that the same user should not book the same shift twice on the same date.