Add functionality to update user availability in admin
The commit introduces new code for enabling an admin to add and update a user's regular availability. This is done by implementing a new test method in the UsersTest.php file, which asserts that the updates to the user availability database table are working as expected. The test supports both addition and modification of user availability records.