Refactor reservation test to use descriptive methods
The complex state definitions have been replaced with more comprehensible method calls in the reservation test. This improves readability and maintenance of the code by making the test setup easier to understand and modify when required. Instead of specifying a maximum of three volunteers with a raw 'state' call, a 'threeVolunteers()' method has been used. Similarly, 'requiresBrother()' and 'allPublishers()' methods have been introduced and used.