Refactor code for user ID validation in ToggleUserOntoShiftReservationController
The validation method for the user id in the ToggleUserOntoShiftReservationController has been streamlined. Rather than checking for the user id and throwing an exception manually, we are now using Laravel's built-in validation function. The 'user' field is now required, must be an integer, and must exist in the users table.