Refactor ToggleShiftReservationControllerRules
The commit removes an unused import (DBPeriod), streamlines the rules, and modifies the formatting of a query method for improved readability. The 'exists:shifts,id' rule was removed as it seemed redundant next to the custom Rule::exists where clause. An additional change is on the array in the Rule::exists, where a trailing comma is added after it. Also, a query method has been formatted better for readability convenience.