Update logic for month and year change in date picker
Refactored the updateMonthYear function in DatePicker.vue to make output more reliable when changing months and years. Instead of creating a new Date with the last day of the previous month, now we're using the lastDayOfMonth function. This will prevent incorrect dates when going from a month with more days to a month with fewer days.